+++ packages/venia-ui/lib/components/ForgotPassword/forgotPassword.js
@@ -1,37 +1,30 @@
-import React, { Fragment, useCallback, useState } from 'react';
+import React, { Fragment } from 'react';
 import { func, shape, string } from 'prop-types';
 
 import { mergeClasses } from '../../classify';
 import ForgotPasswordForm from './ForgotPasswordForm';
 import FormSubmissionSuccessful from './FormSubmissionSuccessful';
 import defaultClasses from './forgotPassword.css';
-import { useUserContext } from '@magento/peregrine/lib/context/user';
+import { useForgotPassword } from '@magento/peregrine/lib/talons/ForgotPassword/useForgotPassword';
 
 const INSTRUCTIONS = 'Enter your email below to receive a password reset link.';
 
 const ForgotPassword = props => {
-    const [{ isResettingPassword }, { resetPassword }] = useUserContext();
-
     const { initialValues, onClose } = props;
 
-    const [inProgress, setInProgress] = useState(false);
-    const [forgotPasswordEmail, setForgotPasswordEmail] = useState(null);
+    const talonProps = useForgotPassword({
+        onClose
+    });
+
+    const {
+        forgotPasswordEmail,
+        handleContinue,
+        handleFormSubmit,
+        inProgress,
+        isResettingPassword
+    } = talonProps;
 
     const classes = mergeClasses(defaultClasses, props.classes);
-
-    const handleFormSubmit = useCallback(
-        async ({ email }) => {
-            setInProgress(true);
-            setForgotPasswordEmail(email);
-            await resetPassword({ email });
-        },
-        [resetPassword]
-    );
-
-    const handleContinue = useCallback(() => {
-        setInProgress(false);
-        onClose();
-    }, [onClose]);
 
     const children = inProgress ? (
         <FormSubmissionSuccessful
